requirements.txt
pyjwt
fastapi
uvicorn
bcrypt
passlib[bcrypt]
python-multipart
